+++ CHANGELOG.md
@@ -5,8 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
-------------
+[0.15.0] - 2024-10-02
+---------------------
 ### Added
 - Added `config.preferSemicolons` which can have a Boolean value to determine whether the example code in the "Usage" section should be terminated by semicolons.
 - Added `config.quoteType` which can have a value of "single" or "double" to determine the type of quotes used in the "Usage" section.
@@ -127,7 +127,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/jbenner-radham/node-readme-md/compare/v0.14.0...HEAD
+[Unreleased]: https://github.com/jbenner-radham/node-readme-md/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.13.1...v0.14.0
 [0.13.1]: https://github.com/jbenner-radham/node-readme-md/compare/v0.13.0...v0.13.1
 [0.13.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.12.0...v0.13.0